+++ src/extension.ts
@@ -1,29 +1,31 @@
 import * as vscode from 'vscode'
 
-export class CodeActionFromCommands implements vscode.CodeActionProvider {
-    constructor(commands: string[]) {
-        this.commands = commands
+const EXTENSION_NAME = 'codeActionFromCommands'
+const CONFIG_NAME = 'codeActions.definitions'
+
+class CodeActionFromCommands implements vscode.CodeActionProvider {
+    constructor(fullName: string, kind: vscode.CodeActionKind) {
+        this.action = new vscode.CodeAction(fullName, kind)
+        this.action.command = { command: fullName, title: fullName }
     }
 
-    public async provideCodeActions(document: vscode.TextDocument, range: vscode.Range | vscode.Selection, context: vscode.CodeActionContext) {
-        if (context.triggerKind === vscode.CodeActionTriggerKind.Automatic) {
-            for (const cmd of this.commands) {
-                await vscode.commands.executeCommand(cmd)
-            }
-        }
-
-        return null
+    public provideCodeActions(document: vscode.TextDocument, range: vscode.Range | vscode.Selection, context: vscode.CodeActionContext) {
+        return context.triggerKind === vscode.CodeActionTriggerKind.Automatic ? [this.action] : null
     }
 
-    readonly commands: string[]
+    readonly action: vscode.CodeAction
 }
 
-export function activate(context: vscode.ExtensionContext) {
-    const extensionName = 'codeActionFromCommands'
-    const definitions = vscode.workspace.getConfiguration(extensionName).get<object>('codeActions.definitions', {})
+function setup(context: vscode.ExtensionContext) {
+    // unregister all previous subscriptions except the one from onDidChangeConfiguration()
+    while (context.subscriptions.length > 1) {
+        context.subscriptions.pop()?.dispose()
+    }
 
-    for (const [codeActionName, codeActionDef] of Object.entries(definitions)) {
-        const languages: string[] = codeActionDef['languages']
+    const config = vscode.workspace.getConfiguration(EXTENSION_NAME).get<object>(CONFIG_NAME, {})
+
+    for (const [codeActionName, codeActionDefinition] of Object.entries(config)) {
+        const languages: string[] = codeActionDefinition['languages']
         let selector: vscode.DocumentSelector
 
         if (!languages) {
@@ -36,15 +38,33 @@ export function activate(context: vscode.ExtensionContext) {
             })
         }
 
-        const providerName = `${extensionName}.${codeActionName}`
-        const commands: string[] = codeActionDef['commands']
-        const provider = new CodeActionFromCommands(commands)
+        const fullName = `${EXTENSION_NAME}.${codeActionName}`
+        const kind = vscode.CodeActionKind.Source.append(fullName)
+        const provider = new CodeActionFromCommands(fullName, kind)
 
         context.subscriptions.push(
             vscode.languages.registerCodeActionsProvider(selector, provider, {
-                providedCodeActionKinds: [vscode.CodeActionKind.Source.append(providerName)]
-            }))
+                providedCodeActionKinds: [kind]
+            }),
+            vscode.commands.registerCommand(fullName, async function () {
+                for (const command of codeActionDefinition['commands']) {
+                    await vscode.commands.executeCommand(command)
+                }
+            }),
+        )
     }
+}
+
+export function activate(context: vscode.ExtensionContext) {
+    context.subscriptions.push(
+        vscode.workspace.onDidChangeConfiguration(evt => {
+            if (evt.affectsConfiguration(`${EXTENSION_NAME}.${CONFIG_NAME}`)) {
+                setup(context)
+            }
+        })
+    )
+
+    setup(context)
 }
 
 export function deactivate() { }